Rewrites the java tool to be modular
