+++ src/main/java/org/opensearch/security/tools/democonfig/ExecutionEnvironment.java
@@ -1,0 +1,9 @@
+package org.opensearch.security.tools.democonfig;
+
+/**
+ * The environment in which the demo config installation script is being executed
+ */
+public enum ExecutionEnvironment {
+    demo, // default value
+    test // to be used only for tests
+}
